+++ odbc-api/src/connection.rs
@@ -454,10 +454,10 @@ impl<'c> Connection<'c> {
     ) -> Result<CursorImpl<StatementImpl<'_>>, Error> {
         let statement = self.allocate_statement()?;
 
-        let catalog_name = catalog_name.map(|s| U16String::from_str(s));
-        let schema_name = schema_name.map(|s| U16String::from_str(s));
-        let table_name = table_name.map(|s| U16String::from_str(s));
-        let table_type = table_type.map(|s| U16String::from_str(s));
+        let catalog_name = catalog_name.map(U16String::from_str);
+        let schema_name = schema_name.map(U16String::from_str);
+        let table_name = table_name.map(U16String::from_str);
+        let table_type = table_type.map(U16String::from_str);
         execute_tables(
             statement,
             catalog_name.as_deref(),